requirements.txt
python-telegram-bot==20.7
requests==2.31.0
python-dotenv==1.0.0
aiohttp==3.9.1
beautifulsoup4==4.12.2
SQLAlchemy==2.0.25 